Decouple parser and semantic by extracting ScopeTracker to Core
