+++ src/core/src/index.ts
@@ -56,6 +56,13 @@ export const Core: CoreNamespace = Object.freeze({
 // Publicly export key AST types at the package root for other packages to
 // import without deep imports. This is the preferred path for type imports
 // across the monorepo.
+export type {
+    GlobalIdentifierTracker,
+    IdentifierRoleManager,
+    ScopeLifecycle,
+    ScopeTracker,
+    ScopeTrackerOptions
+} from "./ast/scope-tracker.js";
 export type { GameMakerAstLocation, GameMakerAstNode, LiteralNode, MutableGameMakerAstNode } from "./ast/types.js";
 export type { DocCommentLines, MutableDocCommentLines } from "./comments/comment-utils.js";
 export type { AbortSignalLike } from "./utils/abort.js";
